[FLINK-37605][runtime] Remove obsolete sink tests
With the removal of SinkV1, all adapter tests have also been testing V2. We can remove the adapter tests and simplify test hierarchy.

(cherry picked from commit fe04a6e9d0f8be33cb9801b4e021c0cd7d0a5b29)
